requirements.txt
torch
torchvision
efficientnet-pytorch
facenet-pytorch
flask
flask-cors
opencv-python
Pillow
numpy
scikit-learn
tqdm
matplotlib
